lto: Make asmlinkage __visible
With LTO the compiler needs to know which function can be
called from assembler, otherwise it would optimize
those functions away. We use the existing asmlinkage
for this, which is already used widely.

Note this causes warnings for static asmlinkage, which
is used in some places. These can be later cleaned up.
static asmlinkage usually makes no sense.

Signed-off-by: Andi Kleen <ak@linux.intel.com>