kbuild: Support padding in kallsyms tables v2
Add support for padding the variable length tables in kallsyms.
This adds a new --pad=XXX option to kallsyms to specify the table lengths,
and another option --pad-file=X to write the table lengths to a file.
Then when a table is shorter than the padding add the necessary padding
at the end.

This allows to replace an existing symbol table later with a different
one that may differ slightly.

Add 5% slack for now in case the prediction is too small
(can happen with LTO)

v2: Fix build error, found by Markus Trippelsdorf.
Signed-off-by: Andi Kleen <ak@linux.intel.com>