x86: Allow disabling HW_BREAKPOINTS and PERF_EVENTS
As suggested by Ingo.

Make HW_BREAKPOINTS a config option. HW_BREAKPOINTS depends on
perf. This allows disabling PERF_EVENTS for systems that don't need it
(e.g. anything not used for development)

Disabling PERF_EVENTS saves over 700k of kernel text (~5% of my
config) and significant data/bss:

   text                data     bss     dec     hex filename
13692640       1922416  1478656 17093712    104d450 obj/vmlinux
12980092       1787544  1470464 16238100     f7c614 obj-noperf/vmlinux

I didn't make it depend on CONFIG_EXPERT for now, as the system should
be very usable even without it. To actually disable perf a couple of
options depending on it need to be disabled, including KVM and
HW_BREAKPOINTS.

Longer term it would be probably nice to have modular perf.

Cc: fweisbec@gmail.com
Cc: peterz@infradead.org
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>