Modify stack depth and size of early buffer
Adding the early_trace_* call adds an entry to the stack frame, so we
don't see the alloc call, so we need to lop less off for early trace
calls.  We'll just have to ignore the extra entry on non-earl calls.

Also, we need a larger buffer for certain events, to avoid drops.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>